Reduce inclusion overhead a little bit
