+++ Code/Framework/AzCore/AzCore/Math/Transform.h
@@ -13,30 +13,9 @@
 #include <AzCore/Math/Vector4.h>
 #include <AzCore/Math/Quaternion.h>
 #include <AzCore/Math/MathUtils.h>
-#include <AzCore/Serialization/SerializeContext.h>
 
 namespace AZ
 {
-    class TransformSerializer
-        : public SerializeContext::IDataSerializer
-    {
-    public:
-        // number of floats in the serialized representation, 4 for rotation, 1 for scale and 3 for translation
-        static constexpr int NumFloats = 8;
-
-        // number of floats in version 1, which used 4 for rotation, 3 for scale and 3 for translation
-        static constexpr int NumFloatsVersion1 = 10;
-
-        // number of floats in version 0, which stored a 3x4 matrix
-        static constexpr int NumFloatsVersion0 = 12;
-
-        size_t Save(const void* classPtr, IO::GenericStream& stream, bool isDataBigEndian) override;
-        size_t DataToText(IO::GenericStream& in, IO::GenericStream& out, bool isDataBigEndian) override;
-        size_t TextToData(const char* text, unsigned int textVersion, IO::GenericStream& stream, bool isDataBigEndian) override;
-        bool Load(void* classPtr, IO::GenericStream& stream, unsigned int version, bool isDataBigEndian) override;
-        bool CompareValueData(const void* lhs, const void* rhs) override;
-    };
-
     //! Limits for transform scale values.
     //! The scale should not be zero to avoid problems with inverting.
     //! @{
