+++ Code/Framework/AzFramework/AzFramework/Visibility/OctreeSystemComponent.cpp
@@ -8,6 +8,7 @@
 
 #include <AzFramework/Visibility/OctreeSystemComponent.h>
 #include <AzCore/Math/ShapeIntersection.h>
+#include <AzCore/Serialization/SerializeContext.h>
 
 namespace AzFramework
 {
