+++ Code/Tools/SceneAPI/SceneCore/Utilities/DebugOutput.h
@@ -11,6 +11,8 @@
 #include <SceneAPI/SceneCore/SceneCoreConfiguration.h>
 #include <SceneAPI/SceneCore/DataTypes/MatrixType.h>
 #include <SceneAPI/SceneCore/Utilities/HashHelper.h>
+#include <AzCore/std/string/string.h>
+#include <AzCore/std/containers/vector.h>
 #include <cinttypes>
 
 namespace AZ::SceneAPI::Utilities
