+++ Gems/AtomLyIntegration/CommonFeatures/Code/Source/PostProcess/RadiusWeightModifier/RadiusWeightModifierComponentController.cpp
@@ -7,6 +7,7 @@
  */
 
 #include <PostProcess/RadiusWeightModifier/RadiusWeightModifierComponentController.h>
+#include <AzCore/RTTI/BehaviorContext.h>
 
 namespace AZ
 {
