+++ Gems/AtomLyIntegration/CommonFeatures/Code/Source/PostProcess/ShapeWeightModifier/ShapeWeightModifierComponentController.cpp
@@ -8,6 +8,7 @@
 
 #include <PostProcess/ShapeWeightModifier/ShapeWeightModifierComponentController.h>
 #include <LmbrCentral/Shape/ShapeComponentBus.h>
+#include <AzCore/Serialization/SerializeContext.h>
 
 namespace AZ
 {
