+++ Gems/AudioSystem/Code/Platform/Common/Default/AudioSystemGemSystemComponent_default.cpp
@@ -9,6 +9,7 @@
 #include <AudioAllocators.h>
 #include <AudioLogger.h>
 #include <SoundCVars.h>
+#include <AzCore/Memory/OSAllocator.h>
 
 namespace Audio::Platform
 {
