+++ Gems/PhysX/Code/Source/System/PhysXJointInterface.cpp
@@ -8,6 +8,7 @@
 
 #include <System/PhysXJointInterface.h>
 #include <AzFramework/Physics/Configuration/JointConfiguration.h>
+#include <AzCore/std/optional.h>
 
 #include <PhysX/Joint/Configuration/PhysXJointConfiguration.h>
 #include <PhysX/Debug/PhysXDebugConfiguration.h>
